[DFT]DFT Regeneration and test for mcp_pmsm_foc_encoder_sam_e70_mchv3_sam_e70_pim MH3-81557

+++ apps/mcp_pmsm_foc_encoder_sam_e70/firmware/src/config/mchv3_sam_e70_pim/peripheral/tc/plib_tc0.h
@@ -96,7 +96,7 @@ void TC0_QuadratureStop (void);
 
 __STATIC_INLINE int16_t TC0_QuadraturePositionGet (void)
 {
-    return (TC0_REGS->TC_CHANNEL[0].TC_CV);
+    return (int16_t)(TC0_REGS->TC_CHANNEL[0].TC_CV);
 }
 
 TC_QUADRATURE_STATUS TC0_QuadratureStatusGet(void);
